excel reader

+++ core/src/main/java/sidorov/common/Element.java
@@ -3,9 +3,9 @@ package sidorov.common;
 public class Element {
     public final int i;
     public final int j;
-    public final int value;
+    public final double value;
 
-    public Element(int i, int j, int value) {
+    public Element(int i, int j, double value) {
         this.i = i;
         this.j = j;
         this.value = value;
